requirements.txt
certifi==2019.3.9
chardet==3.0.4
codegen==1.0
dj-database-url==0.5.0
Django==2.1.1
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-cors-headers==3.0.2
django-favicon==0.1.3
django-heroku==0.3.1
django-mssql==1.8
django-pymssql==1.7.1
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.8
mimeparse==0.1.3
mysql-connector-python==8.0.16
numpy==1.15.4
pandas==0.23.4
patterns==0.3
psycopg2==2.8.2
pymssql==2.1.4
pytz==2019.1
requests==2.21.0
urllib3==1.25.3
whitenoise==4.1.2
